+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 [Migration Guides](https://github.com/urbanairship/android-library/tree/main/documentation/migration)
 
+## Version 16.7.2 September 2, 2022
+
+Patch release that fixes a Message Center data migration and prevents any exceptions with failed migrations going forward.
+
+### Changes
+- Fix crashes due to Message Center database migrations.
+
 ## Version 16.7.1 August 12, 2022
 
 Patch release that prevents potential crashes when downloading files.